Closes #6154: Allow decimal values for cable lengths

+++ netbox/dcim/models/cables.py
@@ -74,7 +74,9 @@ class Cable(PrimaryModel):
     color = ColorField(
         blank=True
     )
-    length = models.PositiveSmallIntegerField(
+    length = models.DecimalField(
+        max_digits=8,
+        decimal_places=2,
         blank=True,
         null=True
     )
